Extract NFTs tests as integration tests
